deployment-templates/function-apps/RedirectAlarmToWebhook/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

requests
azure-functions==1.5.0
